feat: add list item test
