+++ src/popup/components/List/Assignment/ListItem/index.tsx
@@ -27,7 +27,7 @@ function AssignmentListItem({
   };
 
   return (
-    <ListItem disablePadding>
+    <ListItem disablePadding role="listitem">
       {checkable && (
         <ListItemIcon>
           <Checkbox
